dashboard/requirements.txt
pip
python-dotenv
dash
pandas
sqlalchemy
psycopg[binary]
gunicorn